# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Frig-oh Recipe API
  description: OpenAPI specification for the Frig-oh backend (FastAPI). Contains authentication, recipes, image/AI analysis and favorites endpoints.
  version: "1.0.0"
servers:
  - url: http://localhost:8000
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RecipeBase:
      type: object
      properties:
        title:
          type: string
        ingredients:
          type: array
          items:
            type: string
        instructions:
          type: array
          items:
            type: string
        cooking_time:
          type: integer
          description: Cooking time in minutes
        servings:
          type: integer
      required: [title, ingredients, instructions, cooking_time, servings]
    RecipeCreate:
      allOf:
        - $ref: '#/components/schemas/RecipeBase'
    Recipe:
      allOf:
        - $ref: '#/components/schemas/RecipeBase'
        - type: object
          properties:
            _id:
              type: string
            user_id:
              type: string
            created_at:
              type: string
              format: date-time
            is_ai_generated:
              type: boolean
...
    AnalyzeResponse:
      type: object
      properties:
        ingredients:
          type: array
          items:
            type: string
        count:
          type: integer
        message:
          type: string
...
paths:
...
  /recipes/:
    post:
      tags: [Recipes]
      summary: Create a recipe
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeCreate'
      responses:
        '200':
          description: Created recipe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'

    get:
      tags: [Recipes]
      summary: Get user's recipes
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recipe'
  /recipes/generate:
    post:
      tags: [Recipes]
      summary: Generate and save a recipe from a list of ingredients
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
      responses:
        '200':
          description: Generated recipe saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
  /recipes/{recipe_id}:
    get:
      tags: [Recipes]
      summary: Get a single recipe by id
      security:
        - bearerAuth: []
      parameters:
        - name: recipe_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Recipe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '403':
          description: Forbidden
  /recipes/analyze-ingredients:
    post:
      tags: [Recipes]
      summary: Analyze image to detect ingredients
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Ingredients detected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyzeResponse'
        '400':
          description: Bad request (e.g. not an image or too large)
  /recipes/generate-from-photo:
    post:
      tags: [Recipes]
      summary: Analyze photo and generate a recipe (combined flow)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Detected ingredients and saved recipe
          content:
            application/json:
              schema:
                type: object
                properties:
                  detected_ingredients:
                    type: array
                    items:
                      type: string
                  recipe:
                    $ref: '#/components/schemas/Recipe'
                  message:
                    type: string
...